refactor: ws files rework
