+++ package.json
@@ -5,7 +5,7 @@
     "dev": "dotenv -e .env -c -- next dev",
     "build": "next build",
     "initJobs": "JOBS_ENV=1 dotenv -e .env -c -- tsx jobs/initJobs.ts",
-    "initWSServer": "dotenv -e .env -c -- ts-node -O '{\"module\":\"commonjs\"}' -r tsconfig-paths/register ws-service/server.ts",
+    "initWSServer": "dotenv -e .env -c -- ts-node -O '{\"module\":\"commonjs\"}' -r tsconfig-paths/register ws-service/route.ts",
     "start": "cross-env NODE_ENV=production node dist/index.js",
     "prod": "yarn build && NODE_ENV=production next start",
     "test": "dotenv -e .env.test -- ts-node -O '{\"module\":\"commonjs\"}' node_modules/jest/bin/jest.js",
